+++ packages/siwe/scaffold/views/w3m-connecting-siwe-view/index.ts
@@ -9,10 +9,10 @@ import {
   SnackController
 } from '@web3modal/core'
 import { customElement } from '@web3modal/ui'
+import { W3mFrameRpcConstants } from '@web3modal/wallet'
 import { LitElement, html } from 'lit'
 import { state } from 'lit/decorators.js'
 import { SIWEController } from '../../../core/controller/SIWEController.js'
-import { W3mFrameRpcConstants } from '@web3modal/wallet'
 
 @customElement('w3m-connecting-siwe-view')
 export class W3mConnectingSiweView extends LitElement {
@@ -117,7 +117,7 @@ export class W3mConnectingSiweView extends LitElement {
       if (isSmartAccount) {
         SnackController.showError('This application might not support Smart Accounts')
       } else {
-        SnackController.showError('Signature declined')
+        SnackController.showError(error instanceof Error ? error.message : 'Signature declined')
       }
       SIWEController.setStatus('error')