+++ examples/html-ethers5/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @examples/html-ethers5
 
+## 5.0.11-57abd61b5.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @web3modal/ethers5@5.0.11-57abd61b5.0
+
 ## 5.0.10
 
 - chore: update with v5 by @tomiir in https://github.com/WalletConnect/web3modal/pull/2612