+++ packages/ethers/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @web3modal/ethers
 
+## 5.0.11-57abd61b5.0
+
+### Patch Changes
+
+- AppKit Auth canary
+
+- Updated dependencies []:
+  - @web3modal/scaffold-react@5.0.11-57abd61b5.0
+  - @web3modal/scaffold-utils@5.0.11-57abd61b5.0
+  - @web3modal/scaffold-vue@5.0.11-57abd61b5.0
+  - @web3modal/polyfills@5.0.11-57abd61b5.0
+  - @web3modal/scaffold@5.0.11-57abd61b5.0
+  - @web3modal/common@5.0.11-57abd61b5.0
+  - @web3modal/wallet@5.0.11-57abd61b5.0
+  - @web3modal/siwe@5.0.11-57abd61b5.0
+
 ## 5.0.10
 
 - chore: update with v5 by @tomiir in https://github.com/WalletConnect/web3modal/pull/2612