+++ packages/scaffold-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @web3modal/scaffold-react
 
+## 5.0.11-57abd61b5.0
+
+### Patch Changes
+
+- AppKit Auth canary
+
+- Updated dependencies []:
+  - @web3modal/scaffold@5.0.11-57abd61b5.0
+
 ## 5.0.10
 
 - chore: update with v5 by @tomiir in https://github.com/WalletConnect/web3modal/pull/2612